+++ console/src/dataSources/services/bigquery/index.tsx
@@ -213,6 +213,7 @@ export const supportedFeatures: DeepRequired<SupportedFeaturesType> = {
     isolation_level: false,
     connectionSettings: false,
     retries: false,
+    extensions_schema: false,
     pool_timeout: false,
     connection_lifetime: false,
     ssl_certificates: false,